test(entity): Add tests of non-nil Uuid for `user`
